airpods_battery_Show/requirements.txt
pystray>=0.19
Pillow>=9.0
winsdk>=1.0.0b10
